Amélioration de caméra fluide

+++ camera/init.go
@@ -7,7 +7,7 @@ import (
 // Init met en place une caméra.
 func (c *Camera) Init(MapWidth, MapHeight int, AllBlockDisplayed *bool) {
 	if configuration.Global.CameraMode == Static {
-		c.X = configuration.Global.ScreenCenterTileX
-		c.Y = configuration.Global.ScreenCenterTileY
+		c.X = float64(configuration.Global.ScreenCenterTileX)
+		c.Y = float64(configuration.Global.ScreenCenterTileY)
 	}
 }
